refactor(category): change folder name
