+++ FinEdgeFrontend/src/Components/CategoriesPage/CategoriesDataCard.tsx
