+++ FinEdgeFrontend/src/Components/CategoriesPage/CategoryData.tsx
