+++ FinEdgeFrontend/src/Components/CategoriesPage/ExpenditureCategories.tsx
