+++ FinEdgeFrontend/src/Components/CategoriesPage/IncomeCategories.tsx
